+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - feat: Add basic support for attachments #875
+- fix: Crash in SentrySession.serialize #870
 
 ## 6.0.11